finished the screen for settings and drafts
