+++ app/globals.css
@@ -1,3 +1,6 @@
+@import url("https://fonts.googleapis.com/css2?family=Inter:wght@100..900&display=swap");
+@import url("https://fonts.googleapis.com/css2?family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap");
+
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
@@ -52,6 +55,8 @@
     --ring: 222.2 84% 4.9%;
 
     --radius: 0.5rem;
+
+    --poppins: "Poppins", sans-serif;
   }
 
   .dark {
@@ -100,6 +105,10 @@
   .become-a-partner-submit-btn {
     @apply bg-[#E8EDF4] text-black duration-300 px-4 py-1 rounded-sm text-sm;
   }
+}
+
+* {
+  font-family: "Inter", sans-serif;
 }
 
 body {
